refactor: one more pass at naming
and:

- make array elements required (should just be set to empty array so that we don’t always have to check if exists etc)
- make style required
- break INodeExample up into two interfaces
