+++ src/http.ts
@@ -1,23 +1,27 @@
 import { Dictionary } from './basic';
-
-/**
- * Note: inspired by the Axios typings, since that is what Stoplight generally uses under the hood.
- */
+import { IHttpOperation } from './http-spec';
 
 export type HttpMethod = 'get' | 'post' | 'put' | 'patch' | 'delete' | 'head' | 'options';
 export type ExtendedHttpMethod = HttpMethod | 'copy' | 'link' | 'unlink' | 'purge' | 'lock' | 'unlock';
 
+export interface IHttpLog<T = any> {
+  request: IHttpRequest<T>;
+  response?: IHttpResponse<T>;
+  operation?: IHttpOperation;
+}
+
+/** Inspired by the Axios typings, since that is what Stoplight generally uses under the hood. */
 export interface IHttpRequest<T = any> {
+  method: HttpMethod;
+
   /** Can be relative or absolute. If relative, `baseUrl` must also be set. */
   url: string;
-
-  method: HttpMethod;
 
   /** `baseUrl` will be prepended to `url` unless `url` is absolute. */
   baseUrl: string;
 
-  headers?: HttpNameValue;
-  query?: HttpNameValues;
+  headers: HttpNameValue;
+  query: HttpNameValues;
   body?: T;
 }
